Refactor : 판매자 집계 테이블 삭제
주문 완료시 바로 후보가 등록되도록 변경
판매자 집계 테이블은 삭제
