+++ payout/src/main/java/com/mossy/boundedContext/payout/domain/PayoutItem.java
@@ -21,8 +21,14 @@ import static jakarta.persistence.FetchType.LAZY;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class PayoutItem extends BaseIdAndTime {
 
+    // 집계 기능 사용 시 필수
+    // @ManyToOne(fetch = LAZY)
+    // @JoinColumn(name = "payout_id", nullable = false, foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
+    // private Payout payout;
+
+    // 집계 없이 정산만 사용 (nullable로 변경)
     @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "payout_id", nullable = false, foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
+    @JoinColumn(name = "payout_id", foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
     private Payout payout;
 
     @Enumerated(EnumType.STRING)
@@ -71,7 +77,10 @@ public class PayoutItem extends BaseIdAndTime {
 
     private void validatePayoutItem(Payout payout, PayoutEventType eventType, String relTypeCode,
                                     Long relId, LocalDateTime paymentDate, PayoutSeller payee, BigDecimal amount) {
-        if (payout == null) throw new DomainException(ErrorCode.PAYOUT_IS_NULL);
+        // 집계 기능 사용 시 필수
+        // if (payout == null) throw new DomainException(ErrorCode.PAYOUT_IS_NULL);
+
+        // 집계 없이 정산만 사용 시 payout은 optional
         if (eventType == null) throw new DomainException(ErrorCode.PAYOUT_EVENT_TYPE_IS_NULL);
         if (relTypeCode == null || relTypeCode.isBlank()) throw new DomainException(ErrorCode.REL_TYPE_CODE_IS_NULL);
         if (relId == null) throw new DomainException(ErrorCode.REL_ID_IS_NULL);
